[WIP] Move redaction from ruma-signatures to ruma-common
